backend/requirements.txt
aiofiles
aerich
asyncpg==0.29.0
email-validator==2.0.0
fastapi
gunicorn
httpx==0.26.0
passlib[bcrypt]
bcrypt<4.0.0
pydantic-settings
pytest==7.4.4
python-dotenv
python-jose
python-multipart
slowapi
tortoise-orm
tomlkit
tomli_w
uvico